feat(sctp): implement partial reliability chunk dropping
Extracts and properly implements PartialReliabilityBuffer for SCTP in
accordance with TDD principles, replacing a naive list implementation
with a high-performance circular array using Join/Series algebra.

- Creates PartialReliabilityBuffer in commonMain.
- Adds comprehensive failing-then-passing tests in commonTest.
- Cleans up and fixes a flaky coroutine test in SctpReactorSpineTest.
- Verifies changes with jvmTest suite.
