+++ src/commonTest/kotlin/borg/trikeshed/reactor/ngsctp/SctpReactorSpineTest.kt
@@ -53,32 +53,22 @@ class SctpReactorSpineTest {
     @Test
     fun testAssociationScopeCancellation() = runTest {
         val assoc = SctpAssociationScope()
+        var jobStarted = false
         var jobCancelled = false
         val job = assoc.launch {
             try {
-                delay(1000)
+                jobStarted = true
+                delay(10000)
             } finally {
                 jobCancelled = true
             }
         }
+        while (!jobStarted) {
+            delay(10)
+        }
         assoc.close()
         job.join()
         assertTrue(jobCancelled)
-    }
-
-    @Test
-    fun testPartialReliabilityDropsOldestUnacked() {
-        val prBuffer = PartialReliabilityBuffer(capacity = 2)
-        prBuffer.enqueue(1, "chunk1".encodeToByteArray())
-        prBuffer.enqueue(2, "chunk2".encodeToByteArray())
-
-        // At capacity, adding 3 should drop oldest (1)
-        prBuffer.enqueue(3, "chunk3".encodeToByteArray())
-
-        val chunks = prBuffer.getAllUnacked()
-        assertEquals(2, chunks.size)
-        assertEquals(2, chunks[0].first)
-        assertEquals(3, chunks[1].first)
     }
 
     @Test
